refactor(apps/web): move canvas-sync-export and shared types into lib
hooks/useCanvasSync.ts imported canvas-sync-export.ts (a runtime module)
and SyncStatus/UseCanvasSyncOptions (types) back out of hooks/, while
lib/canvas-sync-session.ts imported them the other way — a structural
hooks<->lib import cycle that defeated the point of extracting
canvas-sync-session.ts as a plain non-React leaf module.

Move canvas-sync-export.ts and the shared SyncStatus/UseCanvasSyncOptions/
DirtyEventDetail/dispatchIdentityEvent contract into a new
lib/canvas-sync-types.ts, so every import now flows hooks -> lib only.
hooks/useCanvasSync.ts and hooks/useDirtyState.ts re-export the moved
symbols for existing call sites (DaemonCanvasPage, WorkspaceTopBar).
